+++ packages/frontend/src/features/account/AccountDeletePopperItem.tsx
@@ -4,8 +4,8 @@ import clsx from 'clsx';
 import { FC, useId } from 'react';
 
 import { Drawer } from '$blocks/drawer/drawer';
-import { Button } from '$elements/Button/Button';
-import { ButtonGroup } from '$elements/Button/ButtonGroup';
+import { Button } from '$elements/Button1/Button';
+import { ButtonGroup } from '$elements/Button1/ButtonGroup';
 import { Icon } from '$elements/Icon';
 import { handleAccountDelete } from 'src/actions/account/handleAccountDelete';
