+++ packages/frontend/src/views/OverwriteUserData.tsx
@@ -7,7 +7,7 @@ import { useDispatch } from 'react-redux';
 import { UserDataImportDto } from '$api/generated/financerApi';
 import { DetailsList } from '$blocks/details-list/details-list';
 import { ToastMessageTypes } from '$blocks/toast/toast';
-import { Button } from '$elements/Button/Button';
+import { Button } from '$elements/Button1/Button';
 import { Heading } from '$elements/Heading';
 import { IconName } from '$elements/Icon';
 import {
